Follow advised logging practices
Contrary to what https://docs.python.org/3/howto/logging.html#configuring-logging-for-a-library
advises, we were defining our own StreamHandler with custom formatter and
filter. It is strongly advised for libraries to not do that. We are following that
advice now. In short, the changes done are described below.

- Use module-level loggers with the module name as the logger name. It is described
as a good convention in the https://docs.python.org/3/howto/logging.html#advanced-logging-tutorial.
- Do not add a Handler with custom formatter and filter by default.
- Do not print the latest commit id in the Lifecycle service as it requires reading a file
during the client startup. Instead, the client version is printed. To make test failure
reports easier to follow, the latest commit id is printed for the test logs.
- For tests, a single configuration is added to the root logger in the init file.
That way, all tests will reuse that.
- Do not pass logger extras anymore as they don't make sense without a custom formatter or
filter. We may provide utility implementations for a formatter and a filter and pass
extras even if they won't be used most of the time. I chose not to pass them, but that
can be discussed.
- Remove `logging_level` and `logging_config` confiugration options.
- Update README.