Implement PagingPredicate
This is the port of https://github.com/hazelcast/hazelcast-python-client/pull/212
to the 4.0 branch. However, there are some differences from that PR.

- Documentation of the feature is a little bit different since the client
side sorting is removed.
- Tests are more-or-less ported directly with a few additions for
input validations. However, the above PR was using `six.assertCounEqual`
for assertions which checks equality but not order. For paging predicate
tests, it makes more sense to check ordering.
- Client side codecs and holder classes for PagingPredicates and AnchorDataLists
are added.
- Some public methods, like `get_anchor`, `set_anchor`, `get_nearest_anchor_entry`
is removed. `get_anchor` is defined on the public API on the Java side but I think
it does not provide any functionality to the user.