Convert README format to reStructuredText
We had some problems with the anchor links in our README since
PyPI does not support them when the format is markdown. However,
they work fine with the reStructuredText format. Therefore, README
is converted to that format.

The conversion is mostly done using the `pandoc`. After the
automatic conversion, some problems are fixed manually.