Optimize murmur hash calculation
We were always passing a bytes-like object to the murmur has function
but, we were copying it from the HEAP_DATA_OVERHEAD(8) offset, which
can be huge. I removed that copy and tuned the offset according to our
HEAP_DATA_OVERHEAD.

Also, inlined the _fmix function call and removed the unnecassary
method calls on hash_to_index function.